requirements.txt
dict2obj==1.2.0
Flask==1.0.2
numpy==1.16.0
oandapyV20==0.6.3
omitempty==0.1.1
python-dateutil==2.8.0
requests==2.23.0
SQLAlchemy==1.3.7
TA-Lib==0.4.17